Code Refactoring
* Cleaned up code for readability
* Split up sub-components into their own file for easy access
* Made the ResourcesCard component global for reusability in other sections or pages
